Full service names, semaphores, and updating the update function
The services listed in the service table returned by
connmanctl_parse_services_list is now the full name rather than just
the name less the tech name. UpdateTechnology was removed because it
was a bit redundant compared to the information gathered by calls to
services. The update function now makes an async call to get a list of
services and hands off control to a call back lambda function that
then makes X async calls based on the number of services
returned. When the last async call completes a call to updateMenu is
made. Added a small semaphore structure and associated methods in
order to determine when the update process is complete and when to
make the updateMenu call.