Service List Parsing Redux
The services member is now a flash hash table with the full service
name as the key for that specific service table. Re-wrote unit tests
for parsing the service table as well the the parser itself. Added a
few accessor functions for processing the service table properties
more. Added another helper function to gtable call keys which returns
the list of keys for a given table.